requirements.txt
agate==1.9.1
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asn1crypto==1.5.1
attrs==25.4.0
babel==2.18.0
boto3==1.42.54
botocore==1.42.54
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
daff==1.4.2
dbt-adapters==1.22.6
dbt-common==1.37.2
dbt-core==1.11.6
dbt-extractor==0.6.0
dbt-protos==1.0.431
dbt-semantic-interfaces==0.9.0
dbt-postgres==1.10.0
psycopg2-binary==2.9.10
deepdiff==8.6.1
filelock==3.24.3
idna==3.11
importlib_metadata==8.7.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
keyring==25.7.0
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
minio==7.2.20
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
orderly-set==5.5.0
packaging==26.0
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.9.2
protobuf==6.33.5
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.4
text-unidecode==1.3
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zipp==3.23.0
